backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
opencv-python-headless==4.8.1.78
numpy==1.24.3
paho-mqtt>=2.0.0
bambulabs-api==2.5.8
psutil==5.9.0
python-telegram-bot==13.7
